refactor(memory_profiler): modularize profiler.py into focused components
Split profiler.py (512 lines) into two specialized modules:

New modules:
- profiler_core.py (~320 lines): EndpointProfiler singleton class
  * State management (enabled/disabled, sampling rate)
  * Request counter and smart capture tracking
  * Buffer coordination and auto-flush orchestration
  * Management API and statistics

- decorator.py (~180 lines): profile_endpoint decorator
  * Async/sync endpoint wrappers
  * Request timing and error handling
  * Profile recording coordination
  * Clean separation from core state management

Changes:
- profiler.py: Reduced to thin re-export layer (~20 lines)
- __init__.py: Updated to import from modular structure
- README.md: Documented new modular profiler architecture

Benefits:
- Single responsibility: Each module has one clear purpose
- Better testability: State and decoration logic can be tested independently
- Improved maintainability: Smaller, focused files (~180-320 lines)
- Enhanced readability: Clear separation between concerns
- 100% backward compatible: Existing imports unchanged

Completes the endpoint_profiler modularization alongside the analysis
modules (data_loading, memory_analysis, consumer_analysis, location_analysis,
reporting). The entire endpoint_profiler system is now fully modular with
no files exceeding 320 lines.
