+++ litellm/proxy/common_utils/memory_profiler/endpoint_profiler/__init__.py
@@ -56,8 +56,13 @@ Analysis:
     ...     print(f"Leak detected: {analysis['leak_message']}")
 """
 
-# Main profiler classes
+# Main profiler classes (modular implementation)
+# profiler.py re-exports from profiler_core.py and decorator.py
 from .profiler import EndpointProfiler, profile_endpoint
+
+# Core profiler components (can be imported directly if needed)
+from .profiler_core import EndpointProfiler as _EndpointProfilerCore
+from .decorator import profile_endpoint as _profile_endpoint_decorator
 
 # Storage utilities
 from .storage import ProfileBuffer
@@ -135,7 +140,7 @@ from .constants import (
 )
 
 __all__ = [
-    # Main classes
+    # Main classes (from modular implementation)
     "EndpointProfiler",
     "profile_endpoint",
     "ProfileBuffer",
